refactor(providers): 将 model_context_windows 迁移到 api/providers
