+++ api/providers/model_context_windows.py
